refactor(ansi-to-html): Move more fuzz code to utils
